% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahal_outlier.R
\name{mahal_outlier}
\alias{mahal_outlier}
\title{Mahalanobis Distance Outlier Detection}
\usage{
mahal_outlier(data, p_val_chi = 0.001)
}
\arguments{
\item{p_val_chi}{}
}
\value{

}
\description{
Mahalanobis Distance Outlier Detection
}
