% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoneyData.R
\docType{data}
\name{MoneyData}
\alias{MoneyData}
\title{Money Demand Model for Mexico}
\format{
364 x 5 data frame

\describe{
\item{index}{Monetary Aggregate}
\item{column}
}
}
\usage{
MoneyData
}
\description{
Money Demand Model for Mexico
}
\keyword{datasets}
