% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.R
\name{ols}
\alias{ols}
\title{Ordinary Least Square Estimator}
\usage{
ols(data_ols)
}
\arguments{
\item{data_ols}{This is a function that estimates the OLS estimators and computes
other important statistics}
}
\value{

}
\description{
Ordinary Least Square Estimator
}
