% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_se_white.R
\name{ols_se_white}
\alias{ols_se_white}
\title{OLS with White Robust Standard Errors}
\usage{
ols_se_white(data_ols)
}
\arguments{
\item{data_ols}{}
}
\value{

}
\description{
OLS with White Robust Standard Errors
}
